requirements.txt
django-storages[azure]
python-dotenv==1.0.0
psycopg2-binary==2.9.5
gunicorn==20.1.0
google-cloud-secret-manager==2.15.1
django-environ==0.9.0
django-silk==5.0.3
django-storages==1.13.2
google-cloud-storage==2.7.0
djangorestframework==3.14.0
Pillow==10.1.0
Faker
boto3
gunicorn
django-cors-headers
# git+https://github.com/facebookresearch/detectron2.git
torch==2.2.0
torchvision==0.17
tensorflow
opencv-python
scikit-image
filterpy
pandas
celery==5.4.0
redis==5.0.3